Update appointments patch event listeners

+++ frontend/js/src/appointment.js
@@ -67,15 +67,15 @@ class Appointment{
     deleteAppt = () => { 
         const apptEditButton = document.getElementById(`appt-edit-button-${this.id}`)
         const apptDeleteButton = document.getElementById(`appt-delete-button-${this.id}`)
-        // if (apptDeleteButton.innerText === 'YES') {
+        if (apptDeleteButton.innerText === 'YES') {
             appointmentsAdapter.deleteAppointment(this.id)
-        // }
-        // apptEditButton.innerText = 'Confirm'
-        // apptDeleteButton.innerText = 'YES'
-        // setTimeout(() => {
-        //     apptEditButton.innerText = 'edit'
-        //     apptDeleteButton.innerText = 'delete'
-        // }, 3000)
+        }
+        apptEditButton.innerText = 'Confirm'
+        apptDeleteButton.innerText = 'YES'
+        setTimeout(() => {
+            apptEditButton.innerText = 'edit'
+            apptDeleteButton.innerText = 'delete'
+        }, 3000)
     }
 
     static displayHalfHours = (time) => {
